Algorithms

MetaheuristicsAlgorithms.AEO — Function

AEO(nPop, MaxIt, Low, Up, F_index)

Artificial Ecosystem-based Optimization (AEO) algorithm implementation in Julia.

Arguments:

nPop: Number of individuals in the population.

MaxIt: Maximum number of iterations.

Low: Lower bounds for the search space.

Up: Upper bounds for the search space.

F_index: Function to evaluate the fitness of individuals.

Returns:

AEOResult: A struct containing:

BestF: The best fitness value found.

BestX: The position corresponding to the best fitness.

HisBestFit: A vector of best fitness values at each iteration.

References:

Zhao, Weiguo, Liying Wang, and Zhenxing Zhang. "Artificial ecosystem-based optimization: a novel nature-inspired meta-heuristic algorithm." Neural Computing and Applications 32, no. 13 (2020): 9383-9425.

MetaheuristicsAlgorithms.AFT — Function

AFT(noThieves, itemax, lb, ub, fobj)

The Ali Baba and the Forty Thieves (AFT) algorithm is a meta-heuristic optimization algorithm inspired by the story of Ali Baba and the Forty Thieves. It is designed to solve numerical optimization problems by simulating the behavior of thieves searching for treasures.

Arguments:

noThieves: Number of thieves in the algorithm.

itemax: Maximum number of iterations.

lb: Lower bounds for the search space.

ub: Upper bounds for the search space.

fobj: Objective function to evaluate the fitness of solutions.

Returns:

AFTResult: A struct containing:

fitness: The best fitness value found.

gbest: The position corresponding to the best fitness.

ccurve: A vector of best fitness values at each iteration.

References:

Braik, Malik, Mohammad Hashem Ryalat, and Hussein Al-Zoubi. "A novel meta-heuristic algorithm for solving numerical optimization problems: Ali Baba and the forty thieves." Neural Computing and Applications 34, no. 1 (2022): 409-455.

MetaheuristicsAlgorithms.AHA — Function

AHA(nPop, MaxIt, Low, Up, FunIndex)

The Artificial Hummingbird Algorithm (AHA) is a meta-heuristic optimization algorithm inspired by the foraging behavior of hummingbirds. It is designed to solve numerical optimization problems by simulating the way hummingbirds search for food.

Arguments:

nPop: Number of hummingbirds (population size).

MaxIt: Maximum number of iterations.

Low: Lower bounds for the search space.

Up: Upper bounds for the search space.

FunIndex: Objective function to evaluate the fitness of solutions.

Returns:

AHAResult: A struct containing:

BestF: The best fitness value found.

BestX: The position corresponding to the best fitness.

HisBestFit: A vector of best fitness values at each iteration.

References:

Zhao, Weiguo, Liying Wang, and Seyedali Mirjalili. "Artificial hummingbird algorithm: A new bio-inspired optimizer with its engineering applications." Computer Methods in Applied Mechanics and Engineering 388 (2022): 114194.

MetaheuristicsAlgorithms.ALA — Function

References:

Xiao, Y., Cui, H., Khurma, R. A., & Castillo, P. A. (2025). Artificial lemming algorithm: a novel bionic meta-heuristic technique for solving real-world engineering optimization problems. Artificial Intelligence Review, 58(3), 84.

MetaheuristicsAlgorithms.ALO — Function

References:

Mirjalili, Seyedali. "The ant lion optimizer." Advances in engineering software 83 (2015): 80-98.

MetaheuristicsAlgorithms.AOArithmetic — Function

References:

Abualigah, Laith, Ali Diabat, Seyedali Mirjalili, Mohamed Abd Elaziz, and Amir H. Gandomi. "The arithmetic optimization algorithm." Computer methods in applied mechanics and engineering 376 (2021): 113609.

MetaheuristicsAlgorithms.APO — Function

References:

Wang, Xiaopeng, Václav Snášel, Seyedali Mirjalili, Jeng-Shyang Pan, Lingping Kong, and Hisham A. Shehadeh. "Artificial Protozoa Optimizer (APO): A novel bio-inspired metaheuristic algorithm for engineering optimization." Knowledge-Based Systems 295 (2024): 111737.

MetaheuristicsAlgorithms.ARO — Function

References:

Wang, Liying, Qingjiao Cao, Zhenxing Zhang, Seyedali Mirjalili, and Weiguo Zhao. "Artificial rabbits optimization: A new bio-inspired meta-heuristic algorithm for solving engineering optimization problems." Engineering Applications of Artificial Intelligence 114 (2022): 105082.

MetaheuristicsAlgorithms.ArtemisininO — Function

References:

Yuan, Chong, Dong Zhao, Ali Asghar Heidari, Lei Liu, Yi Chen, Zongda Wu, and Huiling Chen. "Artemisinin optimization based on malaria therapy: Algorithm and applications to medical image segmentation." Displays 84 (2024): 102740.

MetaheuristicsAlgorithms.AVOA — Function

References:

Abdollahzadeh, B., Gharehchopogh, F. S., & Mirjalili, S. (2021). African vultures optimization algorithm: A new nature-inspired metaheuristic algorithm for global optimization problems. Computers & Industrial Engineering, 158, 107408.

MetaheuristicsAlgorithms.BO — Function

References:

Das, Amit Kumar, and Dilip Kumar Pratihar. "Bonobo optimizer (BO): an intelligent heuristic with self-adjusting parameters over continuous spaces and its applications to engineering problems." Applied Intelligence 52, no. 3 (2022): 2942-2974.

MetaheuristicsAlgorithms.BOA — Function

References:

Arora, Sankalap, and Satvir Singh. "Butterfly optimization algorithm: a novel approach for global optimization." Soft computing 23 (2019): 715-734.

MetaheuristicsAlgorithms.CapSA — Function

References:

Braik, Malik, Alaa Sheta, and Heba Al-Hiary. "A novel meta-heuristic search algorithm for solving optimization problems: capuchin search algorithm." Neural computing and applications 33, no. 7 (2021): 2515-2547.

MetaheuristicsAlgorithms.CDO — Function

References:

Shehadeh, Hisham A. "Chernobyl disaster optimizer (CDO): A novel meta-heuristic method for global optimization." Neural Computing and Applications 35, no. 15 (2023): 10733-10749.

MetaheuristicsAlgorithms.ChameleonSA — Function

References:

Braik, Malik Shehadeh. "Chameleon Swarm Algorithm: A bio-inspired optimizer for solving engineering design problems." Expert Systems with Applications 174 (2021): 114685.

MetaheuristicsAlgorithms.ChOA — Function

References:

Khishe, Mohammad, and Mohammad Reza Mosavi. "Chimp optimization algorithm." Expert systems with applications 149 (2020): 113338.

MetaheuristicsAlgorithms.CO — Function

References:

Akbari, Mohammad Amin, Mohsen Zare, Rasoul Azizipanah-Abarghooee, Seyedali Mirjalili, and Mohamed Deriche. "The cheetah optimizer: A nature-inspired metaheuristic algorithm for large-scale optimization problems." Scientific reports 12, no. 1 (2022): 10953.

MetaheuristicsAlgorithms.CoatiOA — Function

References:

Braik, Malik, Alaa Sheta, and Heba Al-Hiary. "A novel meta-heuristic search algorithm for solving optimization problems: capuchin search algorithm." Neural computing and applications 33, no. 7 (2021): 2515-2547.

MetaheuristicsAlgorithms.COOT — Function

References:

Naruei, Iraj, and Farshid Keynia. "A new optimization method based on COOT bird natural life model." Expert Systems with Applications 183 (2021): 115352.

MetaheuristicsAlgorithms.CSBO — Function

References:

Ghasemi, Mojtaba, Mohammad-Amin Akbari, Changhyun Jun, Sayed M. Bateni, Mohsen Zare, Amir Zahedi, Hao-Ting Pai, Shahab S. Band, Massoud Moslehpour, and Kwok-Wing Chau. "Circulatory System Based Optimization (CSBO): an expert multilevel biologically inspired meta-heuristic algorithm." Engineering Applications of Computational Fluid Mechanics 16, no. 1 (2022): 1483-1525.

MetaheuristicsAlgorithms.DBO — Function

References:

Xue J, Shen B. Dung beetle optimizer: A new meta-heuristic algorithm for global optimization. The Journal of Supercomputing. 2023 May;79(7):7305-36.

MetaheuristicsAlgorithms.DDAO — Function

References:

Ghafil, H. N., & Jármai, K. (2020). Dynamic differential annealed optimization: New metaheuristic optimization algorithm for engineering applications. Applied Soft Computing, 93, 106392.

MetaheuristicsAlgorithms.DMOA — Function

References:

Agushaka, Jeffrey O., Absalom E. Ezugwu, and Laith Abualigah. "Dwarf mongoose optimization algorithm." Computer methods in applied mechanics and engineering 391 (2022): 114570.

MetaheuristicsAlgorithms.DO — Function

References:

Zhao, Shijie, Tianran Zhang, Shilin Ma, and Miao Chen. "Dandelion Optimizer: A nature-inspired metaheuristic algorithm for engineering applications." Engineering Applications of Artificial Intelligence 114 (2022): 105075.

MetaheuristicsAlgorithms.DSO — Function

DSO(search_agent, run, LB, UB, dim, ObjFun)

The Deep Sleep Optimiser (DSO) is a human-inspired metaheuristic optimization algorithm that mimics the sleep patterns of humans to find optimal solutions in various optimization problems.

Arguments:

search_agent: Number of search agents (population size).

run: Number of iterations (generations).

LB: Lower bounds of the search space.

UB: Upper bounds of the search space.

ObjFun: Objective function to be minimized.

Returns:

::DSOResult: A struct containing:

Best_Cost: The best cost found during the optimization.

Best_Position: The position corresponding to the best cost.

Best_iteration: A vector of best costs at each iteration.

References:

Oladejo, Sunday O., Stephen O. Ekwe, Lateef A. Akinyemi, and Seyedali A. Mirjalili, The deep sleep optimiser: A human-based metaheuristic approach., IEEE Access (2023).