MetaheuristicsAlgorithms.Individual

MetaheuristicsAlgorithms.Mongoose

MetaheuristicsAlgorithms.individual

MetaheuristicsAlgorithms.AOArithmetic

MetaheuristicsAlgorithms.CDO

MetaheuristicsAlgorithms.ChameleonSA

MetaheuristicsAlgorithms.EO

MetaheuristicsAlgorithms.ETO

MetaheuristicsAlgorithms.ElkHO

MetaheuristicsAlgorithms.Exploration

MetaheuristicsAlgorithms.FATA

MetaheuristicsAlgorithms.GBO

MetaheuristicsAlgorithms.GJO

MetaheuristicsAlgorithms.GKSO

MetaheuristicsAlgorithms.GNDO

MetaheuristicsAlgorithms.GO

MetaheuristicsAlgorithms.GTO

MetaheuristicsAlgorithms.GWO

MetaheuristicsAlgorithms.HBA

MetaheuristicsAlgorithms.HEOA

MetaheuristicsAlgorithms.HGS

MetaheuristicsAlgorithms.Interpolation

MetaheuristicsAlgorithms.JS

MetaheuristicsAlgorithms.LFD

MetaheuristicsAlgorithms.Levy

MetaheuristicsAlgorithms.MPA

MetaheuristicsAlgorithms.MRFO

MetaheuristicsAlgorithms.ParrotO

MetaheuristicsAlgorithms.ROA

MetaheuristicsAlgorithms.RSA

MetaheuristicsAlgorithms.RSO

MetaheuristicsAlgorithms.RUN

MetaheuristicsAlgorithms.RouletteWheelSelection

MetaheuristicsAlgorithms.SCA

MetaheuristicsAlgorithms.SMA

MetaheuristicsAlgorithms.SSA

MetaheuristicsAlgorithms.STOA

MetaheuristicsAlgorithms.SparrowSA

MetaheuristicsAlgorithms.SupplyDO

MetaheuristicsAlgorithms.TSA

MetaheuristicsAlgorithms.WOA

MetaheuristicsAlgorithms.WSO

MetaheuristicsAlgorithms.colleaguesLimitsGenerator

MetaheuristicsAlgorithms.exchange

MetaheuristicsAlgorithms.noh

MetaheuristicsAlgorithms.Individual — Type

Akbari, Mohammad Amin, Mohsen Zare, Rasoul Azizipanah-Abarghooee, Seyedali Mirjalili, and Mohamed Deriche. "The cheetah optimizer: A nature-inspired metaheuristic algorithm for large-scale optimization problems." Scientific reports 12, no. 1 (2022): 10953.

MetaheuristicsAlgorithms.Mongoose — Type

Agushaka, Jeffrey O., Absalom E. Ezugwu, and Laith Abualigah. "Dwarf mongoose optimization algorithm." Computer methods in applied mechanics and engineering 391 (2022): 114570.

MetaheuristicsAlgorithms.individual — Type

Rao, R. Venkata, Vimal J. Savsani, and Dipakkumar P. Vakharia. "Teaching–learning-based optimization: a novel method for constrained mechanical design optimization problems." Computer-aided design 43, no. 3 (2011): 303-315.

MetaheuristicsAlgorithms.AOArithmetic — Method

Abualigah, Laith, Ali Diabat, Seyedali Mirjalili, Mohamed Abd Elaziz, and Amir H. Gandomi. "The arithmetic optimization algorithm." Computer methods in applied mechanics and engineering 376 (2021): 113609.

MetaheuristicsAlgorithms.CDO — Method

Shehadeh, Hisham A. "Chernobyl disaster optimizer (CDO): A novel meta-heuristic method for global optimization." Neural Computing and Applications 35, no. 15 (2023): 10733-10749.

MetaheuristicsAlgorithms.ChameleonSA — Method

Braik, Malik Shehadeh. "Chameleon Swarm Algorithm: A bio-inspired optimizer for solving engineering design problems." Expert Systems with Applications 174 (2021): 114685.

MetaheuristicsAlgorithms.EO — Method

Faramarzi, Afshin, Mohammad Heidarinejad, Brent Stephens, and Seyedali Mirjalili. "Equilibrium optimizer: A novel optimization algorithm." Knowledge-based systems 191 (2020): 105190.

MetaheuristicsAlgorithms.ETO — Method

Luan, Tran Minh, Samir Khatir, Minh Thi Tran, Bernard De Baets, and Thanh Cuong-Le. "Exponential-trigonometric optimization algorithm for solving complicated engineering problems." Computer Methods in Applied Mechanics and Engineering 432 (2024): 117411.

MetaheuristicsAlgorithms.ElkHO — Method

Al-Betar, M.A., Awadallah, M.A., Braik, M.S. et al. Elk herd optimizer: a novel nature-inspired metaheuristic algorithm. Artif Intell Rev 57, 48 (2024). https://doi.org/10.1007/s10462-023-10680-4

MetaheuristicsAlgorithms.Exploration — Method

Abdollahzadeh, Benyamin, Nima Khodadadi, Saeid Barshandeh, Pavel Trojovský, Farhad Soleimanian Gharehchopogh, El-Sayed M. El-kenawy, Laith Abualigah, and Seyedali Mirjalili. "Puma optimizer (PO): A novel metaheuristic optimization algorithm and its application in machine learning." Cluster Computing (2024): 1-49.

MetaheuristicsAlgorithms.FATA — Method

Qi, Ailiang, Dong Zhao, Ali Asghar Heidari, Lei Liu, Yi Chen, and Huiling Chen. "FATA: an efficient optimization method based on geophysics." Neurocomputing 607 (2024): 128289.

MetaheuristicsAlgorithms.GBO — Method

Ahmadianfar, Iman, Omid Bozorg-Haddad, and Xuefeng Chu. "Gradient-based optimizer: A new metaheuristic optimization algorithm." Information Sciences 540 (2020): 131-159.

MetaheuristicsAlgorithms.GJO — Method

Chopra, Nitish, and Muhammad Mohsin Ansari. "Golden jackal optimization: A novel nature-inspired optimizer for engineering applications." Expert Systems with Applications 198 (2022): 116924.

MetaheuristicsAlgorithms.GKSO — Method

Hu, Gang, Yuxuan Guo, Guo Wei, and Laith Abualigah. "Genghis Khan shark optimizer: a novel nature-inspired algorithm for engineering optimization." Advanced Engineering Informatics 58 (2023): 102210.

MetaheuristicsAlgorithms.GNDO — Method

Zhang, Yiying, Zhigang Jin, and Seyedali Mirjalili. "Generalized normal distribution optimization and its applications in parameter extraction of photovoltaic models." Energy Conversion and Management 224 (2020): 113301.

MetaheuristicsAlgorithms.GO — Method

Zhang, Qingke, Hao Gao, Zhi-Hui Zhan, Junqing Li, and Huaxiang Zhang. "Growth Optimizer: A powerful metaheuristic algorithm for solving continuous and discrete global optimization problems." Knowledge-Based Systems 261 (2023): 110206.

MetaheuristicsAlgorithms.GTO — Method

Abdollahzadeh, Benyamin, Farhad Soleimanian Gharehchopogh, and Seyedali Mirjalili. "Artificial gorilla troops optimizer: a new nature‐inspired metaheuristic algorithm for global optimization problems." International Journal of Intelligent Systems 36, no. 10 (2021): 5887-5958.

MetaheuristicsAlgorithms.GWO — Method

Mirjalili, Seyedali, Seyed Mohammad Mirjalili, and Andrew Lewis. "Grey wolf optimizer." Advances in engineering software 69 (2014): 46-61.

MetaheuristicsAlgorithms.HBA — Method

Hashim, Fatma A., Essam H. Houssein, Kashif Hussain, Mai S. Mabrouk, and Walid Al-Atabany. "Honey Badger Algorithm: New metaheuristic algorithm for solving optimization problems." Mathematics and Computers in Simulation 192 (2022): 84-110.

MetaheuristicsAlgorithms.HEOA — Method

Lian, Junbo, and Guohua Hui. "Human evolutionary optimization algorithm." Expert Systems with Applications 241 (2024): 122638.

MetaheuristicsAlgorithms.HGS — Method

Yang, Yutao, Huiling Chen, Ali Asghar Heidari, and Amir H. Gandomi. "Hunger games search: Visions, conception, implementation, deep analysis, perspectives, and towards performance shifts." Expert Systems with Applications 177 (2021): 114864.

MetaheuristicsAlgorithms.Interpolation — Method

Zhao, Weiguo, Liying Wang, Zhenxing Zhang, Seyedali Mirjalili, Nima Khodadadi, and Qiang Ge. "Quadratic Interpolation Optimization (QIO): A new optimization algorithm based on generalized quadratic interpolation and its applications to real-world engineering problems." Computer Methods in Applied Mechanics and Engineering 417 (2023): 116446.

MetaheuristicsAlgorithms.JS — Method

Chou, Jui-Sheng, and Dinh-Nhat Truong. "A novel metaheuristic optimizer inspired by behavior of jellyfish in ocean." Applied Mathematics and Computation 389 (2021): 125535.

MetaheuristicsAlgorithms.LFD — Method

Houssein, Essam H., Mohammed R. Saad, Fatma A. Hashim, Hassan Shaban, and M. Hassaballah. "Lévy flight distribution: A new metaheuristic algorithm for solving engineering optimization problems." Engineering Applications of Artificial Intelligence 94 (2020): 103731.

MetaheuristicsAlgorithms.Levy — Method

Houssein, Essam H., Diego Oliva, Nagwan Abdel Samee, Noha F. Mahmoud, and Marwa M. Emam. "Liver Cancer Algorithm: A novel bio-inspired optimizer." Computers in Biology and Medicine 165 (2023): 107389.

MetaheuristicsAlgorithms.MPA — Method

Faramarzi, Afshin, Mohammad Heidarinejad, Seyedali Mirjalili, and Amir H. Gandomi. "Marine Predators Algorithm: A nature-inspired metaheuristic." Expert systems with applications 152 (2020): 113377.

MetaheuristicsAlgorithms.MRFO — Method

Zhao, Weiguo, Zhenxing Zhang, and Liying Wang. "Manta ray foraging optimization: An effective bio-inspired optimizer for engineering applications." Engineering Applications of Artificial Intelligence 87 (2020): 103300.

MetaheuristicsAlgorithms.ParrotO — Method

Lian, Junbo, Guohua Hui, Ling Ma, Ting Zhu, Xincan Wu, Ali Asghar Heidari, Yi Chen, and Huiling Chen. "Parrot optimizer: Algorithm and applications to medical problems." Computers in Biology and Medicine 172 (2024): 108064.

MetaheuristicsAlgorithms.ROA — Method

Jia, Heming, Xiaoxu Peng, and Chunbo Lang. "Remora optimization algorithm." Expert Systems with Applications 185 (2021): 115665.

MetaheuristicsAlgorithms.RSA — Method

Abualigah, Laith, Mohamed Abd Elaziz, Putra Sumari, Zong Woo Geem, and Amir H. Gandomi. "Reptile Search Algorithm (RSA): A nature-inspired meta-heuristic optimizer." Expert Systems with Applications 191 (2022): 116158.

MetaheuristicsAlgorithms.RSO — Method

Dhiman, Gaurav, Meenakshi Garg, Atulya Nagar, Vijay Kumar, and Mohammad Dehghani. "A novel algorithm for global optimization: rat swarm optimizer." Journal of Ambient Intelligence and Humanized Computing 12 (2021): 8457-8482.

MetaheuristicsAlgorithms.RUN — Method

Ahmadianfar, Iman, Ali Asghar Heidari, Amir H. Gandomi, Xuefeng Chu, and Huiling Chen. "RUN beyond the metaphor: An efficient optimization algorithm based on Runge Kutta method." Expert Systems with Applications 181 (2021): 115079.

MetaheuristicsAlgorithms.RouletteWheelSelection — Method

Mirjalili, Seyedali, Seyed Mohammad Mirjalili, and Abdolreza Hatamlou. "Multi-verse optimizer: a nature-inspired algorithm for global optimization." Neural Computing and Applications 27 (2016): 495-513.

MetaheuristicsAlgorithms.SCA — Method

Mirjalili, Seyedali. "SCA: a sine cosine algorithm for solving optimization problems." Knowledge-based systems 96 (2016): 120-133.

MetaheuristicsAlgorithms.SMA — Method

Li, Shimin, Huiling Chen, Mingjing Wang, Ali Asghar Heidari, and Seyedali Mirjalili. "Slime mould algorithm: A new method for stochastic optimization." Future generation computer systems 111 (2020): 300-323.

MetaheuristicsAlgorithms.SSA — Method

Mirjalili, Seyedali, Amir H. Gandomi, Seyedeh Zahra Mirjalili, Shahrzad Saremi, Hossam Faris, and Seyed Mohammad Mirjalili. "Salp Swarm Algorithm: A bio-inspired optimizer for engineering design problems." Advances in engineering software 114 (2017): 163-191.

MetaheuristicsAlgorithms.STOA — Method

Dhiman, Gaurav, and Amandeep Kaur. "STOA: a bio-inspired based optimization algorithm for industrial engineering problems." Engineering Applications of Artificial Intelligence 82 (2019): 148-174.

MetaheuristicsAlgorithms.SparrowSA — Method

Jiankai Xue & Bo Shen A novel swarm intelligence optimization approach: sparrow search algorithm Systems Science & Control Engineering, 8:1, 22-34, DOI: 10.1080/21642583.2019.1708830 (2020)

MetaheuristicsAlgorithms.SupplyDO — Method

Zhao, Weiguo, Liying Wang, and Zhenxing Zhang. "Supply-demand-based optimization: A novel economics-inspired algorithm for global optimization." Ieee Access 7 (2019): 73182-73206.

MetaheuristicsAlgorithms.TSA — Method

Kaur, Satnam, Lalit K. Awasthi, Amrit Lal Sangal, and Gaurav Dhiman. "Tunicate Swarm Algorithm: A new bio-inspired based metaheuristic paradigm for global optimization." Engineering Applications of Artificial Intelligence 90 (2020): 103541.

MetaheuristicsAlgorithms.WOA — Method

Mirjalili, Seyedali, and Andrew Lewis. "The whale optimization algorithm." Advances in engineering software 95 (2016): 51-67.

MetaheuristicsAlgorithms.WSO — Method

Braik, Malik, Abdelaziz Hammouri, Jaffar Atwan, Mohammed Azmi Al-Betar, and Mohammed A. Awadallah. "White Shark Optimizer: A novel bio-inspired meta-heuristic algorithm for global optimization problems." Knowledge-Based Systems 243 (2022): 108457.

MetaheuristicsAlgorithms.colleaguesLimitsGenerator — Method

Askari, Qamar, Mehreen Saeed, and Irfan Younas. "Heap-based optimizer inspired by corporate rank hierarchy for global optimization." Expert Systems with Applications 161 (2020): 113702.

MetaheuristicsAlgorithms.exchange — Method

Naruei, Iraj, and Farshid Keynia. "Wild horse optimizer: A new meta-heuristic algorithm for solving engineering optimization problems." Engineering with computers 38, no. Suppl 4 (2022): 3025-3056.

MetaheuristicsAlgorithms.noh — Method

Dhiman, Gaurav, and Vijay Kumar. "Spotted hyena optimizer: a novel bio-inspired based metaheuristic technique for engineering applications." Advances in Engineering Software 114 (2017): 48-70.